CHKT-1464 Support dynamic links
