+++ Internal/Debug App/Info.plist
@@ -38,6 +38,7 @@
 		<string>bankid</string>
 		<string>klarna</string>
 		<string>klarnaconsent</string>
+        <string>swish</string>
 	</array>
 	<key>LSRequiresIPhoneOS</key>
 	<true/>
@@ -64,5 +65,7 @@
 	<array>
 		<string>UIInterfaceOrientationPortrait</string>
 	</array>
+    <key>UIUserInterfaceStyle</key>
+    <string>Light</string>
 </dict>
 </plist>
