+++ Sources/PrimerSDK/Classes/Extensions & Utilities/URLExtension.swift
@@ -11,6 +11,10 @@ import Foundation
 
 internal extension URL {
 
+    var isHttpOrHttpsURL: Bool {
+        absoluteString.isHttpOrHttpsURL
+    }
+
     func parseURLQueryParameters() -> [String: String?]? {
         guard let urlComponents = URLComponents(url: self, resolvingAgainstBaseURL: false),
               let queryItems = urlComponents.queryItems else { return nil }
